Move DispatcherProvider to the common package
COSDK-500
