+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/data/remote/api/AnalyticsService.kt
@@ -9,10 +9,10 @@
 package com.adyen.checkout.core.analytics.internal.data.remote.api
 
 import androidx.annotation.RestrictTo
-import com.adyen.checkout.core.DispatcherProvider
 import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsSetupRequest
 import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsSetupResponse
 import com.adyen.checkout.core.analytics.internal.data.remote.model.AnalyticsTrackRequest
+import com.adyen.checkout.core.common.internal.api.DispatcherProvider
 import com.adyen.checkout.core.common.internal.api.HttpClient
 import com.adyen.checkout.core.common.internal.api.post
 import com.adyen.checkout.core.common.internal.model.EmptyResponse
