+++ core/src/main/java/com/adyen/checkout/core/sessions/CheckoutSessionInitializer.kt
@@ -8,8 +8,8 @@
 
 package com.adyen.checkout.core.sessions
 
-import com.adyen.checkout.core.DispatcherProvider
 import com.adyen.checkout.core.Environment
+import com.adyen.checkout.core.common.internal.api.DispatcherProvider
 import com.adyen.checkout.core.common.internal.api.HttpClientFactory
 import com.adyen.checkout.core.components.data.model.Amount
 import com.adyen.checkout.core.sessions.internal.data.api.SessionRepository
